extract controllr class
